+++ share/vss/pedersen/vss_test.go
@@ -127,7 +127,7 @@ func TestVSSWhole(t *testing.T) {
 //}
 
 func TestVSSVerifierNew(t *testing.T) {
-	randIdx := mathRand.Int() % len(verifiersPub)
+	randIdx := uint32(mathRand.Int() % len(verifiersPub))
 	v, err := NewVerifier(suite, verifiersSec[randIdx], dealerPub, verifiersPub)
 	assert.NoError(t, err)
 	assert.Equal(t, randIdx, v.index)
@@ -252,7 +252,7 @@ func TestVSSVerifierReceiveDeal(t *testing.T) {
 	require.NotNil(t, resp)
 	assert.Equal(t, StatusApproval, resp.StatusApproved)
 	assert.Nil(t, err)
-	assert.Equal(t, v.index, int(resp.Index))
+	assert.Equal(t, v.index, resp.Index)
 	assert.Equal(t, dealer.sid, resp.SessionID)
 	assert.Nil(t, schnorr.Verify(suite, v.pub, resp.Hash(suite), resp.Signature))
 	assert.Equal(t, v.responses[uint32(v.index)], resp)
